+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This is the changelog for [OpenAPI Factory](readme.md).
 
 ## 5.0 ##
 * Upgrade to node 12
+* Add greedy matching by add `+` at the end of any dynamic resource replacement.
 
 ## 4.4 ##
 * Provide a way to override the DefaultPathResolver